api/requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.30
pydantic>=2,<3
email-validator>=2
python-jose[cryptography]>=3.3
bcrypt>=4
SQLAlchemy>=2
psycopg2-binary>=2.9
boto3>=1.28
